modifiche classi

+++ backend/src/main/java/com/aether/backend/repository/ColoniaRepository.java
@@ -1,7 +1,7 @@
 package com.aether.backend.repository;
 
 import com.aether.backend.entity.Colonia;
-import org.springframework.data.jpa.repository      .JpaRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColoniaRepository extends JpaRepository<Colonia, Long> {
+public interface ColoniaRepository extends JpaRepository<Colonia, Integer> {
 }
